feat(search): Add layout for search template
